fix: docker setup railwayy
