--- .dockerignore
@@ -1,22 +1,0 @@
-.git
-frontend/
-contracts/
-docs/
-document/
-.agents/
-.opencode/
-.vscode/
-.github/
-**/node_modules/
-**/dist/
-**/.turbo/
-**/.next/
-.env
-.env.*
-!.env.example
-npm-debug.log
-.coverage
-.coverage.*
-.aws
-skills-lock.json
-docker-compose.yml
