+++ backend/package.json
@@ -20,8 +20,12 @@
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.1038.0",
-    "@onecurrency/transactional": "workspace:*",
     "@aws-sdk/s3-request-presigner": "^3.1038.0",
+    "@react-email/render": "^2.0.6",
+    "@react-email/tailwind": "^2.0.7",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-email": "^6.0.0",
     "@hono/standard-validator": "^0.2.2",
     "@neondatabase/serverless": "^1.0.2",
     "@orpc/openapi": "^1.14.0",
@@ -44,6 +48,8 @@
   },
   "devDependencies": {
     "@faker-js/faker": "^10.3.0",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
     "@types/bun": "latest",
     "@types/ws": "^8.18.1",
     "drizzle-kit": "^1.0.0-beta.15-859cf75"
