+++ backend/src/emails/deposit-failed.tsx
@@ -1,16 +1,15 @@
+/** @jsxImportSource react */
 import { Button, Hr, Preview, Text } from "react-email";
 import { EmailLayout } from "./layout";
 
-export type WithdrawalInitiatedEmailProps = {
+export type DepositFailedEmailProps = {
   name: string;
   amountCents: number;
-  withdrawalId: string;
+  depositId: string;
   dashboardUrl: string;
 };
 
 const CENTS_PER_DOLLAR = 100;
-const PAYOUT_BUSINESS_DAYS_MIN = 1;
-const PAYOUT_BUSINESS_DAYS_MAX = 3;
 
 function formatUsd(cents: number): string {
   return new Intl.NumberFormat("en-US", {
@@ -19,61 +18,48 @@ function formatUsd(cents: number): string {
   }).format(cents / CENTS_PER_DOLLAR);
 }
 
-export function WithdrawalInitiatedEmail({
+export function DepositFailedEmail({
   name,
   amountCents,
-  withdrawalId,
+  depositId,
   dashboardUrl,
-}: WithdrawalInitiatedEmailProps) {
+}: DepositFailedEmailProps) {
   return (
     <EmailLayout
-      preview={`Your cash out of ${formatUsd(amountCents)} is on its way`}
+      preview={`We couldn't add ${formatUsd(amountCents)} to your account`}
     >
-      <Preview>{`Your ${formatUsd(amountCents)} cash out is being processed`}</Preview>
-
-      {/* Status indicator */}
-      <Text className="m-0 mb-4 text-center text-4xl">⟳</Text>
-      <Text className="m-0 mb-6 text-center font-semibold text-blue-600 text-sm uppercase tracking-wide">
-        Cash Out Processing
+      <Preview>{`There was a problem adding ${formatUsd(amountCents)} to your OneCurrency account`}</Preview>
+      <Text className="m-0 mb-4 text-center text-4xl">✕</Text>
+      <Text className="m-0 mb-6 text-center font-semibold text-red-500 text-sm uppercase tracking-wide">
+        Add Money Failed
       </Text>
-
-      {/* Amount */}
       <Text className="m-0 mb-2 text-center font-bold text-4xl text-slate-900 tabular-nums">
         {formatUsd(amountCents)}
       </Text>
       <Text className="m-0 mb-8 text-center text-slate-500 text-sm">
-        On its way to your bank account
+        Was not added to your account
       </Text>
-
       <Hr className="my-6 border-slate-200" />
-
-      {/* Message */}
       <Text className="m-0 mb-2 text-base text-slate-900">Hi {name},</Text>
       <Text className="m-0 mb-2 text-base text-slate-700">
-        Your cash out is being processed. Funds typically arrive within{" "}
-        {PAYOUT_BUSINESS_DAYS_MIN}–{PAYOUT_BUSINESS_DAYS_MAX} business days
-        depending on your bank.
+        Your payment was charged, but we encountered a problem crediting your
+        account. Your balance has not been updated.
       </Text>
       <Text className="m-0 mb-6 text-base text-slate-700">
-        You can track the status of your cash out from your account at any time.
+        Please contact our support team at{" "}
+        <a href="mailto:support@onecurrency.tech">support@onecurrency.tech</a>{" "}
+        for assistance or to request a refund.
       </Text>
-
-      {/* CTA */}
       <Button
         className="inline-block rounded-lg bg-blue-600 px-6 py-3 font-semibold text-sm text-white no-underline"
         href={dashboardUrl}
       >
-        View Account
+        Try Again
       </Button>
-
       <Hr className="my-6 border-slate-200" />
-
-      {/* Reference */}
       <Text className="m-0 text-slate-400 text-xs">
-        Receipt reference: {withdrawalId}
+        Receipt reference: {depositId}
       </Text>
     </EmailLayout>
   );
 }
-
-export default WithdrawalInitiatedEmail;
