+++ frontend/package.json
@@ -76,8 +76,7 @@
     "typescript": "^5.9.3",
     "vite": "^8.0.0",
     "vite-tsconfig-paths": "^6.1.1",
-    "vitest": "^4.1.0",
-    "@onecurrency/transactional": "workspace:*"
+    "vitest": "^4.1.0"
   },
   "eslintConfig": {
     "extends": [
