--- packages/transactional/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@onecurrency/transactional",
-  "version": "0.1.0",
-  "private": true,
-  "main": "./dist/src/index.js",
-  "types": "./dist/src/index.d.ts",
-  "scripts": {
-    "build": "tsc --build",
-    "dev": "email dev --dir src/emails",
-    "lint": "biome lint .",
-    "format": "biome format --write ."
-  },
-  "dependencies": {
-    "@react-email/render": "^2.0.6",
-    "@react-email/tailwind": "^2.0.7",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "react-email": "^6.0.0"
-  },
-  "devDependencies": {
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
-    "typescript": "^5.9.3"
-  }
-}
