--- packages/transactional/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "moduleResolution": "bundler",
-    "jsx": "react-jsx",
-    "jsxImportSource": "react",
-    "incremental": true,
-    "rootDir": "./src",
-    "outDir": "dist/src",
-    "noEmit": false,
-    "declaration": true,
-    "declarationMap": true,
-    "composite": true,
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-  },
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["node_modules"]
-}
